bugfix: Cluster Markers didn't always re-render on change

+++ app/src/main/java/me/uni/hiker/ui/screen/map/service/LocationUtils.kt
@@ -244,7 +244,7 @@ fun areRecordedPointsValid(locations: List<RecordedLocation>): Boolean {
     return false
 }
 
-fun clusterTracks(tracks: List<Track>, clusterDistanceThreshold: Double, clusteredTracks: MutableList<AbstractTrack> = mutableListOf()): List<AbstractTrack> {
+fun clusterTracks(tracks: List<Track>, clusterDistanceThreshold: Float, clusteredTracks: MutableList<AbstractTrack> = mutableListOf()): List<AbstractTrack> {
     if (tracks.isEmpty()) return clusteredTracks
 
     tracks.forEach { track ->
